+++ minimalkv/_hstores.py
@@ -6,6 +6,7 @@ from minimalkv.memory import DictStore
 from minimalkv.memory.redisstore import RedisStore
 from minimalkv.net.azurestore import AzureBlockBlobStore
 from minimalkv.net.botostore import BotoStore
+from minimalkv.net.gcstore import GoogleCloudStore
 
 
 class HDictStore(ExtendedKeyspaceMixin, DictStore):
@@ -24,6 +25,10 @@ class HBotoStore(ExtendedKeyspaceMixin, BotoStore):
     def size(self, key):
         k = self.bucket.lookup(self.prefix + key)
         return k.size
+
+
+class HGoogleCloudStore(ExtendedKeyspaceMixin, GoogleCloudStore):
+    pass
 
 
 class HFilesystemStore(ExtendedKeyspaceMixin, FilesystemStore):