+++ docs/changes.rst
@@ -5,6 +5,7 @@ Changelog
 =====
 
 * Add Python 3.6 / 3.9 to build and support matrix.
+* Allow creating ``GoogleCloudStore`` via URL
 * Fix sphinx intersphinx generation and cleanup docs configuration.
 
 1.1.0